Renaming complete from Image2D to Draw2D
